updated the snakemake testing
