+++ config/snakemake/glass.yaml
@@ -1,65 +1,43 @@
-# the two root directories to take the data from and to store the data in
+seed: 42
 source: datasets/glass/resources
-target: datasets/glass/results
+prot_cols: ["AASeq"]
+drug_cols: ["IUPAC"]
 
-
-# structures parameter decides whether to calculate `whole` structure, substructure aligned with `template` or `bsite`.
-# `plddt` means that the loops with low plddt are removed
-# resources/templates has to be present and contain valid structures for last 2 to work
-# plddt works only for alphafold structures, since plddt is written in the b-factor there
-structures: template
-# radius (in Angstroms) around ligands for which residues are considered to be part of the binding site
-bsite:
-  radius: 5
-# radius (in Angstroms) around residues from template to match residues from the query structure
-template:
-  radius: 5
-# cutoff for plddt score to remove flailing loops and whatnot
-plddt:
-  threshold: 60
-
-# how to compute the graphs
-# `rinerator` uses rinerator, also can produce edge labels
-# `distance` just uses a threshold distance between CA atoms
-graph: distance
-
-distance:
-  threshold: 7
-
-# If set to true, only computes the proteins data, skips drugs (for pretraining)
 only_prots: false
 
-# Seed for reproducibility of random processes
-seed: 42
+prots:
+  structs:
+    method: whole # whole, template, bsite, plddt
+    bsite:
+      radius: 5
+    template:
+      radius: 2
+    plddt:
+      threshold: 70
 
-prepare_prots:
-  # can be label, onehot, or esm
-  node_feats: label
-  # can be label, onehot, or none
-  edge_feats: none
+  features:
+    method: rinerator # distance or rinerator
+    distance:
+      threshold: 7
+    node_feats: label # label, onehot, or esm
+    edge_feats: none # label, onehot, or none
 
-prepare_drugs:
+drugs:
   max_num_atoms: 150
-  # can be label, onehot, glycan, or glycanone
-  node_feats: label
-  edge_feats: none
+  node_feats: label # label, onehot, glycan or glycanone
+  edge_feats: none # label, onehot, or none
 
-split:
-  # available methods: random, coldtarget, colddrug
-  method: coldtarget
+split_data:
+  method: random # random, target, drug
   train: 0.7
   val: 0.2
   # test is created as a remainder of train and val
 
 parse_dataset:
-  # filtering can be posneg, balanced or all
-  filtering: posneg
-  # sampling
-  # sampling can be over, under or none
-  sampling: none
-  # classification or regression
-  task: class
+  filtering: all # all, posneg, balanced
+  sampling: none # over, under, none
+  task: class # class, reg
   # threshold values for class problem
-  threshold: 1000
+  threshold: 1.645
   # Whether to apply logarithm with base of 10 to values (for reg, doesn't affect threshold  )
   log: false
